@@ -35,30 +35,32 @@
   },
   "dependencies": {
     "@agm/core": "^1.0.0-beta.6",
-    "@angular/animations": "^7.2.11",
-    "@angular/cdk": "^7.3.6",
-    "@angular/common": "^7.2.11",
-    "@angular/compiler": "^7.2.11",
-    "@angular/core": "^7.2.11",
-    "@angular/forms": "^7.2.11",
-    "@angular/http": "^7.2.11",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "8.0.0",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/http": "^7.2.16",
     "@angular/language-service": "^7.2.11",
-    "@angular/platform-browser": "^7.2.11",
-    "@angular/platform-browser-dynamic": "^7.2.11",
-    "@angular/router": "^7.2.11",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@asymmetrik/ngx-leaflet": "3.0.1",
+    "@fortawesome/fontawesome-free": "^5.1.1",
     "@nebular/auth": "3.4.2",
     "@nebular/bootstrap": "3.4.2",
-    "@nebular/date-fns": "^4.1.0",
+    "@nebular/date-fns": "3.4.2",
     "@nebular/moment": "^4.1.0",
     "@nebular/security": "3.4.2",
     "@nebular/theme": "3.4.2",
     "@ng-bootstrap/ng-bootstrap": "^4.0.0",
-    "@ngx-translate/core": "latest",
-    "@ngx-translate/http-loader": "^4.0.0",
+    "@ngrx/store": "^8.6.0",
+    "@ngx-translate/core": "^10.0.2",
+    "@ngx-translate/http-loader": "^3.0.1",
     "@swimlane/ngx-charts": "^10.0.0",
     "angular-file": "latest",
-    "angular-tree-component": "^7.2.0",
+    "angular-tree-component": "^8.5.6",
     "angular2-chartjs": "0.4.1",
     "angular2-query-builder": "^0.4.2",
     "angular2-toaster": "^7.0.0",
@@ -80,60 +83,66 @@
     "ng2-ckeditor": "^1.2.6",
     "ng2-completer": "2.0.8",
     "ng2-smart-table": "1.3.5",
+    "ng6-file-man": "0.0.10",
     "ngx-dropzone": "^2.1.2",
     "ngx-echarts": "^4.0.1",
+    "ngx-file-drop": "^8.0.8",
     "ngx-lightbox": "^2.1.1",
     "ngx-malihu-scrollbar": "^8.0.0",
+    "ngx-smart-modal": "7.4.1",
     "ngx-summernote": "^0.6.5",
     "ngx-toastr": "^10.0.4",
+    "node-sass": "^4.14.1",
     "normalize.css": "6.0.0",
-    "pace-js": "1.0.2",
+    "pace-js": "^1.0.2",
     "primeicons": "^2.0.0",
     "primeng": "^8.0.3",
     "quill": "^1.3.7",
     "roboto-fontface": "0.8.0",
-    "rxjs": "6.4.0",
+    "rxjs": "^6.4.0",
     "rxjs-compat": "6.3.0",
     "socicon": "3.0.5",
     "summernote": "^0.8.12",
-    "tinymce": "4.5.7",
+    "tinymce": "^4.5.7",
     "tslib": "^1.9.0",
     "typeface-exo": "0.0.22",
     "web-animations-js": "2.2.5",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.12.1",
-    "@angular/cli": "^7.3.7",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/language-service": "7.2.11",
+    "@angular-devkit/build-angular": "~0.803.27",
+    "@angular/cli": "~8.3.27",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^5.4.0",
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
+    "tslib": "^1.13.0",
+    "tslint": "~5.9.1",
+    "typescript": "~3.5.3",
+    "tsickle": "^0.38.1",
     "@compodoc/compodoc": "1.0.1",
-    "@fortawesome/fontawesome-free": "^5.2.0",
     "@types/d3-color": "1.0.5",
     "@types/googlemaps": "^3.30.4",
-    "@types/jasmine": "2.5.54",
-    "@types/jasminewd2": "2.0.3",
     "@types/leaflet": "1.2.3",
-    "@types/node": "6.0.90",
-    "codelyzer": "^4.5.0",
     "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
-    "jasmine-core": "2.6.4",
-    "jasmine-spec-reporter": "4.1.1",
-    "karma": "1.7.1",
-    "karma-chrome-launcher": "2.1.1",
     "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
     "npm-run-all": "4.0.2",
-    "protractor": "5.1.2",
     "rimraf": "2.6.1",
     "stylelint": "7.13.0",
-    "ts-node": "3.2.2",
-    "tslint": "5.7.0",
     "tslint-language-service": "^0.9.9",
-    "typescript": "3.2.4",
     "webpack-bundle-analyzer": "^3.3.2"
   }
 }
