@@ -14,8 +14,8 @@ export const environment = {
   production: false,
   googleApiKey: '',
   mode: 'MARKETPLACE',
-  apiUrl: 'http://15.223.64.6:9090/api',
-  //apiUrl: 'http://aws-demo.shopizer.com:8080/api',
+  // apiUrl: 'http://15.223.64.6:9090/api',
+  apiUrl: 'http://aws-demo.shopizer.com:8080/api',
   client: {
     language: {
       default: 'en',
